Move the platform and openshiftroutes packages under autodetect
