+++ pkg/autodetect/openshiftroutes/main.go
