+++ pkg/autodetect/platform/main.go
